Speed up some functions
For some functions with arity 2, if one of the arguments is not a
Measurement object it was possible to speed up calculations by using
`result' function for functions of arity equal to 1.  This is good for
performance, less good for code beauty.

In addition, bring back degree-related trigonometric functions.  They
aren’t necessary, but having them explicitely defined speeds up
calculations again.