Better explain the formula for uncertainty propagation
Always use "partial derivative" instead of "total derivative", as it may
be a bit misleading (though for a function of independent variables only
they’re exactly the same thing).  Do not introduce the g(x, y, z, ...)
function because it may confuse readers.

The current explanation should be a bit simpler and clearer.