User exposed constructor renamed to measurement
The constructor that user should use has been renamed to `measurement'
instead of `Measurement' in order to follow the Julia convention to have
the lowercase version of type constructor do something more than the
constructor itself.  In particular, here `measurement' is simpler than
`Measurement' but hides to the user all the complexity of `Measurement':
`measurement' takes only the nominal value and the uncertainty as
arguments, and then takes care to create the tag and the derivative
list.