Improve performance of result for multi-arg functions
There was no need to promote all arguments to a common type (which is a costly
operation, due to the creation of new Measurement objects), the only important
thing was to get the common promoted type.  Thus we define a function to
determine the common parameter of all Measurement objects.  This function has
very efficient methods for homogeneous collections (the most frequent case).

The speedup for `sum' function is of one order of magnitutude (tested for
vectors of ~1000 elements).