Fix computation of norms of `AbstractArray{<:Measurement}` (#120)
In Julia v1.9 we currently have

```julia
julia> A = [(14 ± 0.1) (23 ± 0.2); (-12 ± 0.3) (24 ± 0.4)];

julia> norm(A * inv(A) - I, Inf)
0.0 ± 3.5e-19

julia> norm(A * inv(A) - I)
NaN ± NaN
```

This is due to the fact that the Inf-norm has now zero value with non-zero
uncertainty, but some checks in the computation of the p-norms assume that if a
number is non-zero then dividing by it wouldn't result in an overflow (or NaN).
However this is not true for numbers of the type `0 ± u` with `u != 0`, which
aren't the additive identity element but nevertheless would result in
overflow (or NaN) when dividing by them.

While a proper solution would be to have in the language two distinct traits for
"additive identity element" (which is already represented by `iszero`) and
"non-overflowing denominator" (which we currently miss, but I also don't know
any language implementing this trait), Daniel Wennberg pointed out that the
Inf-norm of an array should always have zero-uncertainty if the value is also
zero, thus resulting in a proper "additive identity element" zero which would
satisfy the assumptions in the computations of the other p-norms.  This change
defines the method `LinearAlgebra.normInf(::AbstractArray{<:Measurement})` to
always force zero uncertainty in the Inf-norm when the value is also zero.