Do not extended `Statistics.cor`/`Statistics.cov`
Using the same functions as the `Statistics` ones is a pun, but the functions do
different things, in particular they return scalar values, while the method
defined here return matrices and make it impossible to do the same thing as the
`Statistics` methods.

Also, adding the `Statistics` package as dependency with the sole purpose of
extending methods of functions, which could have been done with an extension,
increases loading time by a factor of 40.